forcer tau_i à être un integer: (int)tau_i; sinon la comparaison t == tau_i(i) impossible

+++ src/genericfun.cpp
@@ -432,7 +432,7 @@ arma::mat matNui(int nD, arma::vec& tau_i, double DeltaT, arma::mat& x0i, arma::
       Mu_t = DeltaT*xi(span(t*nD,(t+1)*nD-1), span(0,n_cols_xi-1))*alpha_mu
       + G_mat_A_0_to_tau_i(span(0,nD-1),span(nD*(t-1),nD*(t-1)+nD-1))*Mu_t;
     }
-    if(t ==tau_i(i)){
+    if(t ==(int)tau_i(i)){
       matNu_i(span(i,i), span(0,nD-1)) = Mu_t.t();
       i++;
     }
